+++ src/bedrock/network/network_system.h
@@ -14,6 +14,7 @@
 
 #pragma once
 
+#include "bedrock/bedrock.h"
 #include "bedrock/core/threading/async.h"
 #include "bedrock/core/utility/binary_stream.h"
 #include "bedrock/forward.h"
@@ -54,6 +55,8 @@ public:
     NetworkConnection *_getConnectionFromId(const NetworkIdentifier &) const;  // Endstone: private -> public
 
 private:
+    ENDSTONE_HOOK virtual bool onNewIncomingConnection(const NetworkIdentifier &id,
+                                                       std::shared_ptr<NetworkPeer> &&peer);  // Endstone: hook
     void _sendInternal(const NetworkIdentifier &id, const Packet &packet, const std::string &data);
     std::unique_ptr<RemoteConnector> remote_connector_;
     std::unique_ptr<ServerLocator> server_locator_;